Include active and modified versions
These are not in table VRSD so they were not being considered. It must be done in a very particular way which I learned by debugging standard function module `SVRS_DISPLAY_DIRECTORY_NEW`.

Let's hope there is no conflict between this active version and the latest version in table VRSD.