Implement filtering
Clicking on a version will display that specific version of that specific object part.

This functionality allows you to see how the code was in a past version.

Because this required re-rendering, the way the page interacts with the handler had to be redesigned. We now have 3 classes: _GUI, _GUI_VIEWER and _GUI_HANDLER.

This was the only missing main functionality. Which means that...

`abapBlame` is now beta!!