Add logic to retrieve old versions
Actually replacing the curent code with an older version.

The logic is working. It was copied from the standard functionality that retrieves a version from the version manager. But it show as a popup to confirm for each part. This is not desirable. I will investigate further and try to find another way to do it which doesn't have popups.

So, this functionality is still not enabled in the HTML page.

Also, the CONVERT TIME STAMP was using the use's time zone. This is fine when wanting to show to the user. But we need the original ones (UTC). For now I'm using UTC for everything. At some point I should use the user's time zone just to show the right time in the UX.