Rewrite summary code
The code is completely rewritten to become more generic, more structured and hopefully more clear and easier to maintain.

Also, transport requests now show to which systems they are already imported (this was actually the reason for the whole rewrite: to make it easy to add this kind of information).